// src/api/models/Provider.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Provider = {
    readonly provider_id: number;
    readonly created_at: string;
    readonly updated_at: string;
    social_name?: string | null;
    birth_datetime?: string | null;
    professional_registration?: number | null;
    specialty_concept?: number | null;
    care_site?: number | null;
};